% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_filename}
\alias{get_filename}
\title{Propose a name of file based on URL}
\usage{
get_filename(url, path = get_gwa_dir(), extension = ".tif")
}
\arguments{
\item{url}{API-url for download}

\item{path}{optional path to a directory for saving the dounloaded file}

\item{extension}{file extension}
}
\value{
filename with path if specified
}
\description{
Propose a name of file based on URL
}
