% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_wind_speed}
\alias{get_wind_speed}
\title{Download wind speed data for a particular country}
\usage{
get_wind_speed(
  country,
  height = 100,
  path = get_gwa_dir(),
  filename = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{country}{3-leter isocode of a country (case sensitive)}

\item{height}{wind turbine hub height: 10, 50, 100, 150, or 200 meters}

\item{path}{optional path to a directory for saving the dounloaded file}

\item{filename}{optional file-name}

\item{overwrite}{logical, should the file be overvriten if exists}
}
\value{
path to the downloaded or existing file
}
\description{
Download wind speed data for a particular country
}
