% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{gwa_get_filename}
\alias{gwa_get_filename}
\title{Propose a name for a downloaded file based on URL}
\usage{
gwa_get_filename(url, path = gwa_get_dir(), extension = ".tif")
}
\arguments{
\item{url}{API-url for download}

\item{path}{optional path to a directory for saving the dounloaded file}

\item{extension}{file extension}
}
\value{
filename with path if specified
}
\description{
Propose a name for a downloaded file based on URL
}
