% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{gwa_get_wind_cf}
\alias{gwa_get_wind_cf}
\title{Download wind capacity factors}
\usage{
gwa_get_wind_cf(
  country,
  IEC = 2,
  path = gwa_get_dir(),
  filename = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{country}{isocode (three letters) of a country (case sensitive)}

\item{IEC}{wind class (1, 2, or 3) according to The International Electrotechnical Commission (IEC)}

\item{path}{optional path to a directory for saving the downloaded file}

\item{filename}{optional file-name}

\item{overwrite}{logical, should the file be overwritten if exists}
}
\value{
path to the downloaded or existing file
}
\description{
Download wind capacity factors
}
