% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{gwa_set_dir}
\alias{gwa_set_dir}
\title{Set directory for/with GWA raster-files}
\usage{
gwa_set_dir(gwa.dir)
}
\arguments{
\item{gwa.dir}{path to a directory where GWA files will downloaded and stored for further use.}
}
\description{
Set directory for/with GWA raster-files
}
