% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalwindatlas-package.R
\docType{package}
\name{globalwindatlas-package}
\alias{globalwindatlas}
\alias{globalwindatlas-package}
\title{globalwindatlas: Access high-resolution (250m) wind potential data in GIS/tif from GlobalWindAtlas.info}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Oleg Lugovoy \email{olugovoy@gmail.com} (\href{https://orcid.org/0000-0001-5546-9875}{ORCID})

}
\keyword{internal}
