% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{gwa_group_locations}
\alias{gwa_group_locations}
\title{(experimentsl) Group locations by intervals of values in given \code{tiff}-file and return geometries of the groups.}
\usage{
gwa_group_locations(
  gwa_tif,
  gis_sf = NULL,
  int = seq(0, 1, by = 0.2),
  eq = "ge",
  simplify = T,
  buffer = T,
  drop_crumps = T,
  verbose = T
)
}
\arguments{
\item{gwa_tif}{tiff-file (terra-object)}

\item{gis_sf}{optional \code{sf} object (map) to crop \code{gwa-tif}}

\item{int}{intervals of values}

\item{eq}{relations to the interval}

\item{simplify}{logical, if TRUE, geometries will be simplified.}

\item{buffer}{logical, if TRUE, buffers will be added to the geometries.}

\item{drop_crumps}{logical, if TRUE, small geometries will be dropped.}

\item{verbose}{logical, should the process be reported.}
}
\value{
\code{sf} object with geometries for each group, defined by intervals.
}
\description{
(experimentsl) Group locations by intervals of values in given \code{tiff}-file and return geometries of the groups.
}
